requirements.txt
pandas
numpy
streamlit
transformers
torchs